requirements.txt
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Caching==2.3.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
redis==6.4.0
six==1.17.0
SQLAlchemy==2.0.41
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
